Fold 'can_reshape' into PhaseValues
Remove the 'can_reshape' argument from Node::Ideal(PhaseGVN *phase, bool
can_reshape) and make it a virtual method of PhaseValues, PhaseGVN, and
PhaseIterGVN instead, to better reflect that 'can_reshape' is a property of the
phase that is running Node::Ideal() (true for PhaseIterGVN, false for PhaseGVN),
as opposed to each individual invocation of Node::Ideal(). Make a few
simplifications based on the invariant can_reshape() == (is_IterGVN() != NULL).